Fixing ActiveAdmin configuration
There are several concepts in this commit:

1) Leaning on the #to_s method of the storage files
2) Updating list page to use abbreviated methods
3) Renaming the :as resource to match how Formtastic was generating
the corresponding form.
4) Exposing attributes for strong parameters

In the case of `as: WorkType`, the underlying Formtastic gem rendered
a the form's object_name as `:work_types_storage`. This meant that
the posted attributes were of the form:

```ruby
{
  work_type_storage: {
    identity: "User input identity"
  }
}
```

ActiveAdmin, however, was expecting the attributes key to be
`:work_type`.

I also needed to ensure that all of the parameters were allowed.